drm: sun4i: hdmi: Fix oops when disabling output
If the display does not indicate audio support in the EDID,
hdmi->hdmi_audio will be NULL. In this case, we avoid setting up audio
output, however we were still attempting to tear it down on disable.

This could lead to a null pointer exception when disabling the output.

Fix the issue by using the same check on tear down as we do on create.

Signed-off-by: Mikel Rychliski <mikel@mikelr.com>